// CodePulseFE/codepulse/src/app/Features/blog-post/models/blog-post.model.ts
import { Category } from "../../Category/models/category.model";

export interface BlogPost {
    id: string;
    title: string;
    shortDescription: string;
    content: string;
    featuredImageUrl: string;
    urlHandle: string;
    publishedDate: Date;
    author: string;
    isVisible: boolean;
    categories: Category[];
}